Updates for the chrome port.  Some parts are working, read desc for details
The active searcher works.  The passive does not.  This is because in
Chrome UDP sockets cannot be bound to an interface/port that is already
in use.  This means that for the vast majority of users the passive
searcher will fail to bind to port 1900.
https://code.google.com/p/chromium/issues/detail?id=246867
In the notifications the device icon cannot be shown as chrome does not
allow applications to reference an external image by url -- it seems to
be possible via data-uri.