updated device id to be the usn which is the uuid + device type. this is unique enough for my router
Which is in violation of the spec as far as I can tell.  But there is
nothing I can do about that.
Also removed the line which didn't update the device if it was the same
response as the last one.  This makes it easier to test, and doesn't
seem to have a performance penalty -- which was the reason for doing it
in the first place.
Lowered the debounce timeout to help with testing.
Fixed xml parsing logic in device factory which was potentially looking
at sub device services and properties to use on the current device --
saw it happening in multiple occasions for the services.